pyproject.toml
[project]
name = "dlect-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain[google-genai,openai]>=0.3.27",
    "langgraph>=0.6.7",
    "langmem>=0.0.29",
    "pydantic>=2.11.9",
    "markdown2>=2.4.0",
    "weasyprint>=60.0",
    "pandas>=2.3.2",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "youtube-transcript-api>=1.2.2",
    "google-api-python-client>=2.182.0",
    "mistralai>=1.9.10",
    "chromadb>=1.1.0",
    "google-genai>=1.38.0",
    "instructor[google-genai]>=1.11.3",
]
